Refactor methods back into helper
